+++ Senparc.Weixin.MP/Senparc.Weixin.MP.Test/Senparc.Weixin.MP.Test.csproj
@@ -71,7 +71,7 @@
     <Compile Include="AdvancedAPIs\CustomService\CustomServiceTest.cs" />
     <Compile Include="AdvancedAPIs\Custom\CustomTest.cs" />
     <Compile Include="AdvancedAPIs\GroupMessage\GroupMessageTest.cs" />
-    <Compile Include="AdvancedAPIs\Groups\CustomTest.cs" />
+    <Compile Include="AdvancedAPIs\Groups\GroupTest.cs" />
     <Compile Include="AdvancedAPIs\Media\MediaTest.cs" />
     <Compile Include="AdvancedAPIs\OAuth\OAuthTest.cs" />
     <Compile Include="AdvancedAPIs\Semantic\SemanticTest.cs" />
